Add: Makefile and main.cpp

+++ Assignment2/Assignment2a/header.h
@@ -6,9 +6,9 @@
 
 namespace Assignment2a
 {
-    std::pair<char*, int> file_to_array(const std::string& path);
-    void reverse_array(char* const buffer, const int size);
-    void array_to_file(char* const buffer, const int size);
+    std::pair<char*, uintmax_t> file_to_array(const std::string& path);
+    void reverse_array(char* const buffer, const uintmax_t size);
+    void array_to_file(char* const buffer, const uintmax_t size);
 }
